Give stories better names, extract DnD stories
